+++ sip/execution_control/docker_api/sip_docker_swarm/docker_swarm_client.py
@@ -421,10 +421,10 @@ class DockerSwarmClient:
         Args:
             service_id (str): docker swarm service id
 
-        Returns,
+        Returns:
             str, replicated level of the service
-        """
 
+        """
         # Raise an exception if we are not a manager
         if not self._manager:
             raise RuntimeError('Only the Swarm manager node can retrieve '
@@ -441,8 +441,9 @@ class DockerSwarmClient:
         Args:
             service_id (str): docker swarm service id
 
-        Returns,
+        Returns:
             str, replication level of the service
+
         """
         # Initialising empty list
         replicas = []
@@ -520,6 +521,9 @@ class DockerSwarmClient:
             if 'logging' in key:
                 self._parse_logging(value, service_config)
                 service_config.pop('logging')
+            if 'environment' in key:
+                service_config['env'] = value
+                service_config.pop('environment')
 
         return service_config