+++ sip/execution_control/processing_block_controller/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.1] - 2019-01-23
+### Changed
+- Updated to Docker Swarm API 1.0.10.
+
 ## [1.3.0] - 2019-01-09
 ### Changed
 - Updated dockerfile to run the PBC as a root user. While not generally a good