+++ sip/execution_control/processing_controller/CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on
 and this project adheres to
  [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.7] - 2019-01-23
+### Changed
+- Updated to Docker Swarm API 1.0.10.
 
 ## [1.2.6] - 2019-01-10
 ### Changed